refactor: separate folder per bot
